// extensions/cactus-plugin-htlc-coordinator/src/main/typescript/generated/openapi/typescript-axios/api.ts
export interface CounterpartyHTLCRequest {
    /**
     * 
     * @type {HtlcPackage}
     * @memberof CounterpartyHTLCRequest
     */
    htlcPackage: HtlcPackage;
    /**
     * connector Instance Id for the connector plugin
     * @type {string}
     * @memberof CounterpartyHTLCRequest
     */
    connectorInstanceId: string;
    /**
     * keychainId for the keychain plugin
     * @type {string}
     * @memberof CounterpartyHTLCRequest
     */
    keychainId: string;
    /**
     * Id for the HTLC
     * @type {string}
     * @memberof CounterpartyHTLCRequest
     */
    htlcId: string;
    /**
     * 
     * @type {Web3SigningCredential}
     * @memberof CounterpartyHTLCRequest
     */
    web3SigningCredential: Web3SigningCredential;
    /**
     * 
     * @type {number}
     * @memberof CounterpartyHTLCRequest
     */
    gas?: number;
}
/**
 * 
 * @export
 * @enum {string}
 */

export enum HtlcPackage {
    Besu = 'BESU',
    BesuErc20 = 'BESU_ERC20'
}

/**
 * 
 * @export
 * @interface OwnHTLCRequest
 */
export interface OwnHTLCRequest {
    /**
     * 
     * @type {HtlcPackage}
     * @memberof OwnHTLCRequest
     */
    htlcPackage: HtlcPackage;
    /**
     * connector Instance Id for the connector plugin
     * @type {string}
     * @memberof OwnHTLCRequest
     */
    connectorInstanceId: string;
    /**
     * keychainId for the keychain plugin
     * @type {string}
     * @memberof OwnHTLCRequest
     */
    keychainId: string;
    /**
     * 
     * @type {Array<any>}
     * @memberof OwnHTLCRequest
     */
    constructorArgs: Array<any>;
    /**
     * 
     * @type {Web3SigningCredential}
     * @memberof OwnHTLCRequest
     */
    web3SigningCredential: Web3SigningCredential;
    /**
     * Input amount to lock
     * @type {number}
     * @memberof OwnHTLCRequest
     */
    inputAmount: number;
    /**
     * Output amount to lock
     * @type {number}
     * @memberof OwnHTLCRequest
     */
    outputAmount: number;
    /**
     * Timestamp to expire the contract
     * @type {number}
     * @memberof OwnHTLCRequest
     */
    expiration: number;
    /**
     * Hashlock needed to refund the amount
     * @type {string}
     * @memberof OwnHTLCRequest
     */
    hashLock: string;
    /**
     * The token address
     * @type {string}
     * @memberof OwnHTLCRequest
     */
    tokenAddress: string;
    /**
     * The receiver address
     * @type {string}
     * @memberof OwnHTLCRequest
     */
    receiver: string;
    /**
     * The output network id
     * @type {string}
     * @memberof OwnHTLCRequest
     */
    outputNetwork: string;
    /**
     * The output addreess to receive the tokens
     * @type {string}
     * @memberof OwnHTLCRequest
     */
    outputAddress: string;
    /**
     * 
     * @type {number}
     * @memberof OwnHTLCRequest
     */
    gas?: number;
}
/**
 * @type Web3SigningCredential
 * @export
 */
export type Web3SigningCredential = Web3SigningCredentialCactusKeychainRef | Web3SigningCredentialNone | Web3SigningCredentialPrivateKeyHex;

/**
 * 
 * @export
 * @interface Web3SigningCredentialCactusKeychainRef
 */
export interface Web3SigningCredentialCactusKeychainRef {
    /**
     * 
     * @type {Web3SigningCredentialType}
     * @memberof Web3SigningCredentialCactusKeychainRef
     */
    type: Web3SigningCredentialType;
    /**
     * The ethereum account (public key) that the credential  belongs to. Basically the username in the traditional  terminology of authentication.
     * @type {string}
     * @memberof Web3SigningCredentialCactusKeychainRef
     */
    ethAccount: string;
    /**
     * The key to use when looking up the the keychain entry holding the secret pointed to by the  keychainEntryKey parameter.
     * @type {string}
     * @memberof Web3SigningCredentialCactusKeychainRef
     */
    keychainEntryKey: string;
    /**
     * The keychain ID to use when looking up the the keychain plugin instance that will be used to retrieve the secret pointed to by the keychainEntryKey parameter.
     * @type {string}
     * @memberof Web3SigningCredentialCactusKeychainRef
     */
    keychainId: string;
}
/**
 * Using this denotes that there is no signing required because the transaction is pre-signed.
 * @export
 * @interface Web3SigningCredentialNone
 */
export interface Web3SigningCredentialNone {
    /**
     * 
     * @type {Web3SigningCredentialType}
     * @memberof Web3SigningCredentialNone
     */
    type: Web3SigningCredentialType;
}
/**
 * 
 * @export
 * @interface Web3SigningCredentialPrivateKeyHex
 */
export interface Web3SigningCredentialPrivateKeyHex {
    /**
     * 
     * @type {Web3SigningCredentialType}
     * @memberof Web3SigningCredentialPrivateKeyHex
     */
    type: Web3SigningCredentialType;
    /**
     * The ethereum account (public key) that the credential belongs to. Basically the username in the traditional terminology of authentication.
     * @type {string}
     * @memberof Web3SigningCredentialPrivateKeyHex
     */
    ethAccount: string;
    /**
     * The HEX encoded private key of an eth account.
     * @type {string}
     * @memberof Web3SigningCredentialPrivateKeyHex
     */
    secret: string;
}
/**
 * 
 * @export
 * @enum {string}
 */

export enum Web3SigningCredentialType {
    CactusKeychainRef = 'CACTUS_KEYCHAIN_REF',
    GethKeychainPassword = 'GETH_KEYCHAIN_PASSWORD',
    PrivateKeyHex = 'PRIVATE_KEY_HEX',
    None = 'NONE'
}

/**
 * 
 * @export
 * @interface WithdrawCounterpartyRequest
 */
export interface WithdrawCounterpartyRequest {
    /**
     * 
     * @type {HtlcPackage}
     * @memberof WithdrawCounterpartyRequest
     */
    htlcPackage: HtlcPackage;
    /**
     * connector Instance Id for the connector plugin
     * @type {string}
     * @memberof WithdrawCounterpartyRequest
     */
    connectorInstanceId: string;
    /**
     * keychainId for the keychain plugin
     * @type {string}
     * @memberof WithdrawCounterpartyRequest
     */
    keychainId: string;
    /**
     * contractId for the contract
     * @type {string}
     * @memberof WithdrawCounterpartyRequest
     */
    contractId?: string;
    /**
     * 
     * @type {Web3SigningCredential}
     * @memberof WithdrawCounterpartyRequest
     */
    web3SigningCredential: Web3SigningCredential;
    /**
     * Id for the HTLC
     * @type {string}
     * @memberof WithdrawCounterpartyRequest
     */
    htlcId: string;
    /**
     * Counterparty HTLC secret
     * @type {string}
     * @memberof WithdrawCounterpartyRequest
     */
    secret: string;
    /**
     * 
     * @type {number}
     * @memberof WithdrawCounterpartyRequest
     */
    gas?: number;
}